+++ CHANGELOG.md
@@ -538,6 +538,12 @@ $$Pow({x}, {n})=Pow\left({x}^2, \dfrac{n}{2}\right) * Pow\left(x, n \mod 2\right
 
 - Recursion
 
+## 0116 - Populating Next Right Pointers in Each Node
+
+- Recursion
+- Connect right of left subtree to the left of right subtree
+- Continue recursivly to left and right subtree, util tree is nil
+
 ## 0118 - Pascal's Triangle
 
 - Nothing